Relocate keccak::encode to eth_abi
