+++ shared/src/types/mod.rs
@@ -3,6 +3,7 @@
 pub mod address;
 pub mod chain;
 pub mod dylib;
+pub mod eth_abi;
 pub mod eth_bridge_pool;
 pub mod ethereum_events;
 pub mod governance;
